backend/pyproject.toml
[project]
name = "calendar-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "caldav>=1.4",
    "icalendar>=6.0",
    "httpx>=0.27",
    "python-dateutil>=2.9",
    "sse-starlette>=2.1",
    "holidays>=0.58",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
